pyproject.toml
[project]
name = "comp-of-ts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "websockets>=12.0",
    "python-decouple>=3.8"
]
